Reorganize crate structure
