+++ Cargo.toml
@@ -1,6 +1,9 @@
 [workspace]
 resolver = "3"
-members = ["valid/colamd"]
+members = [
+    "crates/sparse_colamd",
+    "crates/sparse_colamd/valid",
+]
 
 [package]
 name = "sparse"
@@ -14,7 +17,7 @@ categories = ["mathematics"]
 readme = "README.md"
 
 [workspace.package]
-version = "0.5.0"
+version = "0.5.1"
 description = "Sparse"
 repository = "https://github.com/lokyhark/sparse"
 license = "MIT"
@@ -24,6 +27,8 @@ readme = "README.md"
 
 
 [dependencies]
+sparse_array = { version = "0.5.1", path = "crates/sparse_array" }
+sparse_colamd = { version = "0.5.1", path = "crates/sparse_colamd" }
 
 [lints.clippy]
 all = "warn"
