+++ crates/sparse_array/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "sparse_array"
+edition = "2024"
+version.workspace = true
+description.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+readme.workspace = true
+
+[dependencies]
