+++ crates/sparse_colamd/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "colamd"
+name = "sparse_colamd"
 edition = "2024"
 version.workspace = true
 description.workspace = true
@@ -10,9 +10,4 @@ categories.workspace = true
 readme.workspace = true
 
 [dependencies]
-sparse = { path = "../.." }
-flate2 = "1.1.9"
-tar = "0.4.46"
-
-[build-dependencies]
-cc = "1.0"
+sparse_array = { version = "*", path = "../sparse_array" }
