+++ crates/sparse_colamd/ref/License.txt
