+++ crates/sparse_colamd/ref/colamd.h
