+++ crates/sparse_colamd/src/int.rs
