+++ crates/sparse_colamd/src/row.rs
