+++ crates/sparse_colamd/src/score.rs
