+++ crates/sparse_colamd/src/size.rs
