+++ crates/sparse_colamd/valid/summary.md
