--- valid/colamd/build.rs
@@ -1,5 +1,0 @@
-fn main() {
-    println!("cargo::rerun-if-changed=ref/colamd.h");
-    println!("cargo::rerun-if-changed=ref/colamd.c");
-    cc::Build::new().include("ref").file("ref/colamd.c").flags(["-std=c99", "-O2", "-Wall", "-Wextra"]).compile("reference");
-}
